requirements.txt
pytest
PyYAML
pydoit
attrs
XlsxWriter
